fixed issue with responsivness on landing page, signup, and login
