+++ client/src/pages/Profile.jsx
@@ -1,14 +1,6 @@
-// import React from 'react';
-// import { Navigate, useParams } from 'react-router-dom';
-// import { useQuery } from '@apollo/client';
-
-// import { QUERY_USER, QUERY_ME } from '../utils/queries';
-
-// import Auth from '../utils/auth';
-
 const Profile = () => {
   return (
-    <div>
+    <div className="container">
       <div className="flex-row justify-center mb-3">
         <h2 className="col-12 col-md-10 bg-dark text-light p-3 mb-5">
           Viewing your profile.
