--- client/src/styles/Landing.css
@@ -1,70 +1,0 @@
-.main {
-  margin-top: 70px;
-    height: 100vh;
-    width:100%;
-    background-color: white;
-   
-    
-     
-  }
-  
-  .main h1 {
-    font-size: 100px;
-    text-transform: uppercase;
-    font-weight: 800;
-    text-align: center;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    z-index: 9;
-    background-image: linear-gradient(to top, rgb(60, 140, 214),rgb(13, 148, 65),rgb(182, 46, 46),rgb(39, 125, 165),green,brown,rgb(144, 202, 144));
-    -webkit-background-clip: text; /* For Safari */
-    background-clip: text;
-    color: transparent; /* Hide the original text color */
-  
-  
-  }
-  .main button {
-  background-color:rgb(184, 64, 35);
-  font-size: 25px;
-  padding: 20px 30px;
-  margin: 0 20px 20px 0;
-  border-radius: 50px;
-  z-index: 9;
-  color: aliceblue;
-  font-weight: 800;
-  }
-  
-  .main h1:nth-child(2) {
-   
-    margin-top: 60px;
-    z-index: 9;
-  }
-  .main .left-img {
-    height: 160px;
-    position: absolute;
-    top: 8%;
-    left: 11%;
-    border-radius: 30px
-  }
-  .main .right-img {
-    height: 200px;
-    position: absolute;
-    bottom: 28%;
-    right: 6%;
-    border-radius: 30px
-  }
-  .main .top-img {
-    height: 120px;
-    position: absolute;
-    top: 1%;
-    right: 24%;
-  
-  }
-  .main .bottom-img {
-    height: 280px;
-    position: absolute;
-    bottom: 3%;
-    left: 14%;
-    border-radius: 30px
-  }
