+++ client/src/styles/Login.css
@@ -5,7 +5,7 @@
 }
 
 body {
-  background-color: #F4F7F5;
+  background-color: white;
 }
 
 .box {
